move dynesty plotter to nest plotter
