+++ autofit/non_linear/plot/nest_plotters.py
@@ -1,13 +1,11 @@
 from dynesty import plotting as dyplot
 from functools import wraps
-import numpy as np
 
-from autofit.plot import SamplesPlotter
-
+from autofit.non_linear.plot import SamplesPlotter
 from autofit.non_linear.plot.samples_plotters import skip_plot_in_test_mode
 
-def log_value_error(func):
 
+def log_value_error(func):
     @wraps(func)
     def wrapper(self, *args, **kwargs):
         """
@@ -42,14 +40,15 @@ def log_value_error(func):
 
     return wrapper
 
-class DynestyPlotter(SamplesPlotter):
+
+class NestPlotter(SamplesPlotter):
 
     @skip_plot_in_test_mode
     @log_value_error
     def boundplot(self, **kwargs):
         """
         Plots the in-built ``dynesty`` plot ``boundplot``.
-        
+
         This figure plots the bounding distribution used to propose either (1) live points
         at a given iteration or (2) a specific dead point during
         the course of a run, projected onto the two dimensions specified
@@ -57,7 +56,6 @@ class DynestyPlotter(SamplesPlotter):
         """
 
         dyplot.boundplot(
-            dims=self.model.total_free_parameters,
             results=self.samples.search_internal.results,
             labels=self.model.parameter_labels_with_superscripts_latex,
             **kwargs
@@ -77,56 +75,31 @@ class DynestyPlotter(SamplesPlotter):
         the course of a run, projected onto all pairs of dimensions.
         """
 
-        import os
-
-        if os.environ.get("PYAUTOFIT_TEST_MODE") == "1":
-            return
-
-        import corner
-
-        corner.corner(
-            data=np.asarray(self.samples.parameter_lists),
-            weight_list=self.samples.weight_list,
+        dyplot.cornerbound(
+            results=self.samples.search_internal.results,
             labels=self.model.parameter_labels_with_superscripts_latex,
             **kwargs
         )
 
-        self.output.to_figure(structure=None, auto_filename="corner")
+        self.output.to_figure(structure=None, auto_filename="cornerbound")
         self.close()
-
-        # dyplot.cornerbound(
-        #     results=self.samples.search_internal.results,
-        #     labels=self.model.parameter_labels_with_superscripts_latex,
-        #     **kwargs
-        # )
-        #
-        # self.output.to_figure(structure=None, auto_filename="cornerbound")
-        # self.close()
 
     @skip_plot_in_test_mode
     @log_value_error
-    def corner(self, **kwargs):
+    def cornerplot(self, **kwargs):
         """
-        Plots the in-built ``dynesty`` plot ``corner``.
+        Plots the in-built ``dynesty`` plot ``cornerplot``.
 
         This figure plots a corner plot of the 1-D and 2-D marginalized posteriors.
         """
 
-        import os
-
-        if os.environ.get("PYAUTOFIT_TEST_MODE") == "1":
-            return
-
-        import corner
-
-        corner.corner(
-            data=np.asarray(self.samples.parameter_lists),
-            weight_list=self.samples.weight_list,
+        dyplot.cornerplot(
+            results=self.samples.search_internal.results,
             labels=self.model.parameter_labels_with_superscripts_latex,
             **kwargs
         )
 
-        self.output.to_figure(structure=None, auto_filename="corner")
+        self.output.to_figure(structure=None, auto_filename="cornerplot")
         self.close()
 
     @skip_plot_in_test_mode
